[WIP, RISCV] Add inline asm constraint A for RISC-V
Summary:
This allows the constraint A to be used in inline asm for RISC-V, which allows an address held in a register to be used.

This patch adds the minimal amount of code required to get operands with the right constraints to compile. Semantically it would be nice to be able to use `setAllowsRegister` rather than `setAllowsMemory` and fall through to the backend to verify and lower the register as a memory operand.

Reviewers: asb

Subscribers: eraman, rbar, johnrusso, simoncook, apazos, sabuasal, niosHD, kito-cheng, shiva0217, jrtc27, zzheng, edward-jones, mgrang, rogfer01, MartinMosbeck, brucehoult, the_o, rkruppe, PkmX, jocewei, cfe-commits

Differential Revision: https://reviews.llvm.org/D54295